Define actions as const
